more on the structure
