+++ src/init.c
@@ -1,0 +1,138 @@
+#include "init.h"
+
+void RCC_Configuration(void){
+	/* Enable DMA and GPIOA Clocks */
+	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_DMA1 | RCC_AHB1Periph_GPIOA | RCC_AHB1Periph_GPIOA, ENABLE);
+	/* Enable DAC1 and TIM6 clocks */
+	RCC_APB1PeriphClockCmd(RCC_APB1Periph_DAC | RCC_APB1Periph_TIM6, ENABLE);
+
+	RCC_APB2PeriphClockCmd(RCC_APB2Periph_SYSCFG, ENABLE);
+}
+
+void NVIC_Configuration(void)
+{
+	NVIC_InitTypeDef NVIC_InitStructure;
+	/* Enable and set EXTI Line0 Interrupt to the lowest priority */
+	NVIC_InitStructure.NVIC_IRQChannel = EXTI0_IRQn | EXTI1_IRQn | EXTI2_IRQn \
+			| EXTI3_IRQn | EXTI4_IRQn ;// |EXTI9_5_IRQn | EXTI15_10_IRQn;
+	NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 0x01;
+	NVIC_InitStructure.NVIC_IRQChannelSubPriority = 0x01;
+	NVIC_InitStructure.NVIC_IRQChannelCmd = ENABLE;
+	NVIC_Init(&NVIC_InitStructure);
+}
+
+void GPIO_Configuration(void)
+{
+	GPIO_InitTypeDef GPIO_InitStruct;
+	GPIO_InitTypeDef GPIO_InitStructure;
+	/* Pack the struct */
+	GPIO_InitStruct.GPIO_Pin = GPIO_Pin_4;
+	GPIO_InitStruct.GPIO_Speed = GPIO_Speed_50MHz;
+	GPIO_InitStruct.GPIO_OType = GPIO_OType_PP;
+	GPIO_InitStruct.GPIO_PuPd = GPIO_PuPd_NOPULL;
+	/* Call Init function */
+	GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+	// GPIO_InitStructure.GPIO_Mode = GPIO_Mode_IN;
+	// GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_NOPULL;
+	// GPIO_InitStructure.GPIO_Pin = GPIO_Pin_0 | GPIO_Pin_1 | GPIO_Pin_2 | GPIO_Pin_3 \
+	// 	  | GPIO_Pin_4 ;| GPIO_Pin_5 | GPIO_Pin_6 | GPIO_Pin_7 | GPIO_Pin_8 \
+	// 	  | GPIO_Pin_9 | GPIO_Pin_10 | GPIO_Pin_11 | GPIO_Pin_12 | GPIO_Pin_13 \
+	// 	  | GPIO_Pin_14 | GPIO_Pin_15;
+	// GPIO_Init(GPIOB, &GPIO_InitStructure);
+}
+void DMA_Configuration(uint16_t* wavBuffer)
+{
+	DMA_InitTypeDef DMA_InitStructure;
+	//Initialize the structure to default values
+	DMA_StructInit(&DMA_InitStructure);
+	DMA_InitStructure.DMA_Channel = DMA_Channel_7;
+	DMA_InitStructure.DMA_PeripheralBaseAddr = (uint32_t)(DAC_BASE + 0x08);  //DAC channel1 12-bit right-aligned data holding register (ref manual pg. 264)
+	DMA_InitStructure.DMA_Memory0BaseAddr = (uint32_t)wavBuffer;
+	DMA_InitStructure.DMA_DIR = DMA_DIR_MemoryToPeripheral;
+	DMA_InitStructure.DMA_BufferSize = DACBUFFERSIZE;
+	DMA_InitStructure.DMA_PeripheralInc = DMA_PeripheralInc_Disable;
+	DMA_InitStructure.DMA_MemoryInc = DMA_MemoryInc_Enable;
+	DMA_InitStructure.DMA_PeripheralDataSize = DMA_PeripheralDataSize_HalfWord;
+	DMA_InitStructure.DMA_MemoryDataSize = DMA_MemoryDataSize_HalfWord;
+	DMA_InitStructure.DMA_Mode = DMA_Mode_Circular;
+	DMA_InitStructure.DMA_Priority = DMA_Priority_High;
+	DMA_InitStructure.DMA_FIFOMode = DMA_FIFOMode_Disable;
+	DMA_InitStructure.DMA_FIFOThreshold = DMA_FIFOThreshold_Full;
+	DMA_InitStructure.DMA_MemoryBurst = DMA_MemoryBurst_Single;
+	DMA_InitStructure.DMA_PeripheralBurst = DMA_PeripheralBurst_Single;
+	/* Call Init function */
+	DMA_Init(DMA1_Stream5, &DMA_InitStructure);
+	/* Enable DMA */
+	DMA_Cmd(DMA1_Stream5, ENABLE);
+}
+/**
+  * @brief  Configures the Timers.
+  * @param  wavePeriod (period of timer), preScaler (prescaler for timer)
+  * @retval : None
+  */
+void Timer_Configuration(uint16_t wavPeriod, uint16_t preScaler)
+{
+	TIM_TimeBaseInitTypeDef  TIM_TimeBaseStruct;
+	/* pack Timer struct */
+	TIM_TimeBaseStruct.TIM_Period = wavPeriod-1;
+	TIM_TimeBaseStruct.TIM_Prescaler = preScaler-1;
+	TIM_TimeBaseStruct.TIM_ClockDivision = TIM_CKD_DIV1;
+	TIM_TimeBaseStruct.TIM_CounterMode = TIM_CounterMode_Up;
+	TIM_TimeBaseStruct.TIM_RepetitionCounter = 0x0000;
+	/* Call init function */
+	TIM_TimeBaseInit(TIM6, &TIM_TimeBaseStruct);
+	/* Select Timer to trigger DAC */
+	TIM_SelectOutputTrigger(TIM6, TIM_TRGOSource_Update);
+	/* TIM6 enable counter */
+	TIM_Cmd(TIM6, ENABLE);
+}
+/**
+  * @brief  Configures the DAC
+  * @param  None
+  * @retval : None
+  */
+void DAC_Configuration(void)
+{
+	DAC_InitTypeDef DAC_InitStruct;
+	DAC_InitStruct.DAC_Trigger = DAC_Trigger_T6_TRGO;
+	DAC_InitStruct.DAC_WaveGeneration = DAC_WaveGeneration_None;
+	DAC_InitStruct.DAC_LFSRUnmask_TriangleAmplitude = DAC_LFSRUnmask_Bit0;
+	DAC_InitStruct.DAC_OutputBuffer = DAC_OutputBuffer_Enable;
+	DAC_Init(DAC_Channel_1, &DAC_InitStruct);
+	DAC_DMACmd(DAC_Channel_1, ENABLE);
+	/* Enable DAC Channel1: Once the DAC channel1 is enabled, PA.04 is automatically connected to the DAC converter. */
+	DAC_Cmd(DAC_Channel_1, ENABLE);
+}
+
+void EXTILine_Config(uint32_t GPIO_Pin, uint32_t EXTI_Lin, uint32_t EXTI_PinSource, uint32_t EXTI_IRQn)
+{
+  GPIO_InitTypeDef   GPIO_InitStructure;
+  NVIC_InitTypeDef   NVIC_InitStructure;
+  EXTI_InitTypeDef EXTI_InitStructure;
+  
+  /* Configure PA0 pin as input floating */
+  GPIO_InitStructure.GPIO_Mode = GPIO_Mode_IN;
+  GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_NOPULL;
+  // GPIO_InitStructure.GPIO_OType = GPIO_OType_PP;
+  GPIO_InitStructure.GPIO_Speed = GPIO_Speed_100MHz;
+  GPIO_InitStructure.GPIO_Pin = GPIO_Pin;
+  GPIO_Init(GPIOA, &GPIO_InitStructure);
+
+  /* Connect EXTI Line0 to PA0 pin */
+  SYSCFG_EXTILineConfig(EXTI_PortSourceGPIOA, EXTI_PinSource);
+
+  /* Configure EXTI Line0 */
+  EXTI_InitStructure.EXTI_Line = EXTI_Lin;// | EXTI_Line1 | EXTI_Line2 | EXTI_Line3 ;
+  EXTI_InitStructure.EXTI_Mode = EXTI_Mode_Interrupt;
+  EXTI_InitStructure.EXTI_Trigger = EXTI_Trigger_Rising;  
+  EXTI_InitStructure.EXTI_LineCmd = ENABLE;
+  EXTI_Init(&EXTI_InitStructure);
+
+  /* Enable and set EXTI Line0 Interrupt to the lowest priority */
+  NVIC_InitStructure.NVIC_IRQChannel = EXTI_IRQn;
+  NVIC_InitStructure.NVIC_IRQChannelPreemptionPriority = 0x01;
+  NVIC_InitStructure.NVIC_IRQChannelSubPriority = 0x01;
+  NVIC_InitStructure.NVIC_IRQChannelCmd = ENABLE;
+  NVIC_Init(&NVIC_InitStructure);
+}
