+++ src/main.h
@@ -9,13 +9,17 @@
 #define MAIN_H_
 
 /* Includes */
+#include <math.h>
+
 #include "stm32f4xx.h"
 #include "stm32f4_discovery.h"
-#include <math.h>
+#include "init.h"
 #include "pitch.h"
+#include "wavetable.h"
+#include "lcd.h"
+
 
 /* Private macro */
-#define DACBUFFERSIZE 		250
 #define TIMER6_PRESCALER	2 	/* produces a 42MHz tick */
 #define TIMER_CLOCK			84E6 /* TIM6 runs at 84MHz */
 
@@ -25,19 +29,7 @@ uint32_t buttonVal = 0;
 
 //uint32_t dma_mode = DMA_Mode_Circular;
 uint16_t volume = 10;
-uint16_t SineWaveBuffer[DACBUFFERSIZE]; 	/* Array for  waveform 1*/
-uint16_t SawtoothBuffer[DACBUFFERSIZE]; 	/* Array for  waveform 2*/
-uint16_t PulseWaveTable[DACBUFFERSIZE];
-uint16_t TriangleWaveTable[DACBUFFERSIZE];
 
-
-void RCC_Configuration(void);
-void DMA_Configuration( uint16_t* wavBuffer );
-void NVIC_Configuration(void);
-void GPIO_Configuration(void);
-void UART_Configuration(void);
-void Timer_Configuration(uint16_t wavPeriod, uint16_t preScaler);
-void DAC_Configuration(void);
 // void EXTILines_Config(void);
 void EXTILine_Config(uint32_t GPIO_Pin, uint32_t EXTI_Lin, uint32_t EXTI_PinSource, uint32_t EXTI_IRQn);
 void delay_ms(uint32_t milli);
