--- src/pitch.c
@@ -1,6 +1,0 @@
-/*
- * pitch.c
- *
- *  Created on: Sep 22, 2015
- *      Author: Poon Donohoe
- */
